Split code for api usage
